Milestone: Resolved errors and conformed to new way of generating documents
